+++ Subsystems/Storage/StorageSubsystem.cs
@@ -1,6 +1,7 @@
 using BepInEx.Configuration;
 using HarmonyLib;
 using ServerSync;
+using UnityEngine;
 using Wonderland.Core;
 
 namespace Wonderland.Subsystems.Storage
@@ -18,10 +19,12 @@ namespace Wonderland.Subsystems.Storage
         public void OnWorldReady()
         {
             ItemCache.Initialize();
+            ContainerRowsEngine.Initialize();
         }
 
         public void OnUpdate()
         {
+            ContainerRowsEngine.OnUpdate(Time.deltaTime);
         }
 
         public void Shutdown()